+++ bpy_classes.py
@@ -101,6 +101,68 @@ classes = [
     modules.material.sub_matl_data.SUB_PG_matl_vertex_attribute,
     modules.material.sub_matl_data.SUB_PG_matl_linked_material,
     modules.material.sub_matl_data.SUB_PG_sub_matl_data,
+    modules.swing.sub_swing_data.SUB_PG_armature_swing_bone,
+    modules.swing.sub_swing_data.SUB_PG_armature_swing_bone_children,
+    modules.swing.sub_swing_data.SUB_PG_swing_bone_collision,
+    modules.swing.sub_swing_data.SUB_PG_swing_bone,
+    modules.swing.sub_swing_data.SUB_PG_swing_bone_chain,
+    modules.swing.sub_swing_data.SUB_PG_swing_sphere,
+    modules.swing.sub_swing_data.SUB_PG_swing_oval,
+    modules.swing.sub_swing_data.SUB_PG_swing_ellipsoid,
+    modules.swing.sub_swing_data.SUB_PG_swing_capsule,
+    modules.swing.sub_swing_data.SUB_PG_swing_plane,
+    modules.swing.sub_swing_data.SUB_PG_swing_connection,
+    modules.swing.sub_swing_data.SUB_PG_sub_swing_data,
+    modules.swing.sub_swing_data.SUB_PG_blender_bone_data,
+    modules.swing.ui.SUB_PT_swing_io,
+    modules.swing.ui.SUB_PT_active_bone_swing_info,
+    modules.swing.ui.SUB_PT_swing_data_master,
+    modules.swing.ui.SUB_PT_swing_bone_chains,
+    modules.swing.ui.SUB_UL_swing_bone_chains,
+    modules.swing.ui.SUB_UL_swing_bones,
+    modules.swing.ui.SUB_UL_swing_bone_collisions,
+    modules.swing.ui.SUB_MT_swing_bone_collision_add,
+    modules.swing.ui.SUB_PT_swing_data_spheres,
+    modules.swing.ui.SUB_UL_swing_data_spheres,
+    modules.swing.ui.SUB_MT_swing_data_spheres_context_menu,
+    modules.swing.ui.SUB_PT_swing_data_ovals,
+    modules.swing.ui.SUB_UL_swing_data_ovals,
+    modules.swing.ui.SUB_MT_swing_data_ovals_context_menu,
+    modules.swing.ui.SUB_PT_swing_data_ellipsoids,
+    modules.swing.ui.SUB_UL_swing_data_ellipsoids,
+    modules.swing.ui.SUB_MT_swing_data_ellipsoids_context_menu,
+    modules.swing.ui.SUB_PT_swing_data_capsules,
+    modules.swing.ui.SUB_UL_swing_data_capsules,
+    modules.swing.ui.SUB_MT_swing_data_capsules_context_menu,
+    modules.swing.ui.SUB_PT_swing_data_planes,
+    modules.swing.ui.SUB_UL_swing_data_planes,
+    modules.swing.ui.SUB_MT_swing_data_planes_context_menu,
+    modules.swing.ui.SUB_PT_swing_data_connections,
+    modules.swing.ui.SUB_UL_swing_data_connections,
+    modules.swing.ui.SUB_MT_swing_data_connections_context_menu,
+    modules.swing.operators.SUB_OP_swing_bone_chain_add,
+    modules.swing.operators.SUB_OP_swing_bone_chain_remove,
+    modules.swing.operators.SUB_OP_swing_bone_chain_length_edit,
+    modules.swing.operators.SUB_OP_swing_bone_collision_add_sphere,
+    modules.swing.operators.SUB_OP_swing_bone_collision_add_oval,
+    modules.swing.operators.SUB_OP_swing_bone_collision_add_ellipsoid,
+    modules.swing.operators.SUB_OP_swing_bone_collision_add_capsule,
+    modules.swing.operators.SUB_OP_swing_bone_collision_add_plane,
+    modules.swing.operators.SUB_OP_swing_bone_collision_remove,
+    modules.swing.operators.SUB_OP_swing_data_sphere_add,
+    modules.swing.operators.SUB_OP_swing_data_sphere_remove,
+    modules.swing.operators.SUB_OP_swing_data_oval_add,
+    modules.swing.operators.SUB_OP_swing_data_oval_remove,
+    modules.swing.operators.SUB_OP_swing_data_ellipsoid_add,
+    modules.swing.operators.SUB_OP_swing_data_ellipsoid_remove,
+    modules.swing.operators.SUB_OP_swing_data_capsule_add,
+    modules.swing.operators.SUB_OP_swing_data_capsule_remove,
+    modules.swing.operators.SUB_OP_swing_data_plane_add,
+    modules.swing.operators.SUB_OP_swing_data_plane_remove,
+    modules.swing.operators.SUB_OP_swing_data_connection_add,
+    modules.swing.operators.SUB_OP_swing_data_connection_remove,
+    modules.swing.operators.SUB_OP_swing_import,
+    modules.swing.operators.SUB_OP_swing_export,
     operators.eye_material_custom_vector_31_modal.SUB_OP_eye_material_custom_vector_31_modal,
     properties.SubSceneProperties,
     shader_nodes.custom_sampler_node.SUB_CSN_ultimate_sampler,
